Support setFormat across both sharp and gm.
This patch allows the output format to be specified as a string for
both these engines. This commit also somewhat standardises on the
specification of output type via "setFormat" - convert named formats
On the query string for sharp while in the case of gm remove a wad
Of data shuffling code which is no longer required given it’s native
support of the operation.