Carefully revise error handling and stream teardown.
Between the duplex stream of the pipeline and hijackresponse there
was a situation causing an errant end event to cancel the streams
just prior to fully handling an error error event. This is somehow
reminiscent of another situation where having bidirectional streams
in the mix created confusion.

Attempt to get around this by using pipe() in only one direction
and in the other handing the events by hand.