Raise minimum node version to 16.
Also add 18 and 20 to the test matrix. This addresses some confusion
around the currently supported versions of node - namely, a commit
that stopped testing the library on 14 ended up being shipped in a
path release as opposed to a major. This shouldn't actually affect
anything because the underpinnings (impro) did not change anything
on the support front, but take the opportunity to nail a stake in
the ground. Jump ahead of impro to 16 in order to leave some extra
wiggle room to iterate on things there.