Call time on testing images by asserting on their resemblance.
This has caused no end of maintenance issues due to their internal
use of the canvas module which is itself a native dependency. Not
only has this caused upgrade issues but these "close within reason"
assertions themselves have needed adjusting thus buys us nothing.

Declare bankruptcy on the maintenance burden instead follow impro
in sidestepping issues by checking the files in and removing use
of unexpected-resemble.

Nudge impro to its next patch version to allow speciyfing the use
of imagemagick as an engine option on the pipeline in the tests.